Update agent to work with new size of tags
Depends-On: I6cc6685644dd42b2afb007788001488394eb9b44
Depends-On: Iaa767c8b4146c282fa66dc0e1ff645f137ca6963

+++ src/vnsw/agent/cmn/agent.h
@@ -199,7 +199,7 @@ void intrusive_ptr_add_ref(const SecurityLoggingObject *p);
 
 //class SecurityGroup;
 typedef std::vector<int> SecurityGroupList;
-typedef std::vector<int> TagList;
+typedef std::vector<uint64_t> TagList;
 typedef std::vector<boost::uuids::uuid> UuidList;
 typedef std::vector<std::string> CommunityList;
